pyproject.toml
[project]
name = "ecaa"
version = "0.2.0"
description = "Engineering Cognitive AI Agents"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "chainlit>=1.1.301",
    "openai>=1.34.0",
    "litellm>=1.40.14",
    "python-frontmatter>=1.1.0",
    "loguru>=0.7.2",
    "chromadb>=0.5.5",
    "sentence-transformers>=3.0.1",
    "pytest-asyncio>=0.24.0",
    "newspaper3k>=0.2.8",
    "lxml[html_clean]>=5.3.0",
    "pydantic>=2.9.2",
]
readme = "README.md"
requires-python = ">=3.12.4"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.7.1",
    "ipykernel>=6.29.5",
    "pytest>=8.3.2",
    "ruff>=0.6.5",
    "pip>=24.2",
]

[tool.ruff]
line-length = 55 # Book formatting
fix = true

[tool.basedpyright]
typeCheckingMode = "strict"
verboseOutput = true
venvPath = ".venv"
pythonVersion = "3.12"

[tool.rye.scripts]
precommit = "pre-commit run --all-files"
ch02_basic = "chainlit run src/ch02/winston_basic.py"
ch02_history = "chainlit run src/ch02/winston_history.py"
ch02_streaming = "chainlit run src/ch02/winston_streaming.py"
ch02_tool = "chainlit run src/ch02/winston_tool.py"
ch03_greeting = "chainlit run src/ch03/winston_greeting.py"
ch03_advanced = "chainlit run src/ch03/winston_advanced.py"
ch03_router = "chainlit run src/ch03/winston_router.py"
ch04_knows = "chainlit run src/ch04/winston_knows.py"
ch05_converses = "chainlit run src/ch05/winston_converses.py"
ch05_ponders = "chainlit run src/ch05/winston_ponders.py"
ch05_remembers = "chainlit run src/ch05/winston_remembers.py"
donald_goals = "chainlit run --port 8003 src/ch07/donald_goals.py -w"
donald_qa = "chainlit run --port 8003 src/ch07/donald_qa.py -w"
